+++ src/test/java/io/github/kirstenali/deepj/activations/ActivationsTest.java
@@ -12,57 +12,57 @@ public class ActivationsTest {
     @Test
     void relu_forwardBackward() {
         ReLU relu = new ReLU();
-        Tensor x = Tensor.from2D(new double[][]{
+        Tensor x = Tensor.from2D(new float[][]{
                 {-1, 0, 2},
                 {3, -4, 5}
         });
 
         Tensor y = relu.forward(x);
-        TestSupport.assertTensorAllClose(y, Tensor.from2D(new double[][]{
+        TestSupport.assertTensorAllClose(y, Tensor.from2D(new float[][]{
                 {0, 0, 2},
                 {3, 0, 5}
-        }), 1e-12);
+        }), 1e-12f);
 
-        Tensor gradOut = Tensor.from2D(new double[][]{
+        Tensor gradOut = Tensor.from2D(new float[][]{
                 {1, 1, 1},
                 {2, 2, 2}
         });
         Tensor gx = relu.backward(gradOut);
-        TestSupport.assertTensorAllClose(gx, Tensor.from2D(new double[][]{
+        TestSupport.assertTensorAllClose(gx, Tensor.from2D(new float[][]{
                 {0, 0, 1},
                 {2, 0, 2}
-        }), 1e-12);
+        }), 1e-12f);
     }
 
     @Test
     void sigmoid_outputsInRange_andBackwardNonNegativeForPositiveUpstream() {
         Sigmoid s = new Sigmoid();
-        Tensor x = Tensor.from2D(new double[][]{{-10, 0, 10}});
+        Tensor x = Tensor.from2D(new float[][]{{-10, 0, 10}});
         Tensor y = s.forward(x);
 
         for (int c = 0; c < y.cols; c++) {
-            Assertions.assertTrue(y.data[c] > 0.0 && y.data[c] < 1.0);
+            Assertions.assertTrue(y.data[c] > 0.0f && y.data[c] < 1.0f);
         }
 
-        Tensor gradOut = Tensor.from2D(new double[][]{{1, 1, 1}});
+        Tensor gradOut = Tensor.from2D(new float[][]{{1, 1, 1}});
         Tensor gx = s.backward(gradOut);
         for (int c = 0; c < gx.cols; c++) {
-            Assertions.assertTrue(gx.data[c] >= 0.0, "sigmoid' should be >= 0");
+            Assertions.assertTrue(gx.data[c] >= 0.0f, "sigmoid' should be >= 0");
         }
     }
 
     @Test
     void tanh_isOddFunctionApprox() {
         Tanh t = new Tanh();
-        Tensor x1 = Tensor.from2D(new double[][]{{0.5, -0.5}});
+        Tensor x1 = Tensor.from2D(new float[][]{{0.5f, -0.5f}});
         Tensor y1 = t.forward(x1);
-        Assertions.assertEquals(y1.data[0], -y1.data[1], 1e-12);
+        Assertions.assertEquals(y1.data[0], -y1.data[1], 1e-12f);
     }
 
     @Test
     void gelu_isSmooth_andMonotonicAroundZero() {
         GELU g = new GELU();
-        Tensor x = Tensor.from2D(new double[][]{{-1e-3, 0.0, 1e-3}});
+        Tensor x = Tensor.from2D(new float[][]{{-1e-3f, 0.0f, 1e-3f}});
         Tensor y = g.forward(x);
 
         Assertions.assertTrue(y.data[0] < y.data[1]);
@@ -72,16 +72,16 @@ public class ActivationsTest {
     @Test
     void softmax_rowsSumTo1() {
         Softmax sm = new Softmax();
-        Tensor logits = Tensor.from2D(new double[][]{
+        Tensor logits = Tensor.from2D(new float[][]{
                 {1, 2, 3},
                 {-1, 0, 1}
         });
         Tensor p = sm.forward(logits);
 
         for (int r = 0; r < p.rows; r++) {
-            double sum = 0.0;
+            float sum = 0.0f;
             for (int c = 0; c < p.cols; c++) sum += p.data[r * p.cols + c];
-            Assertions.assertEquals(1.0, sum, 1e-6);
+            Assertions.assertEquals(1.0f, sum, 1e-6f);
         }
 
         // backward should require forward
@@ -91,7 +91,7 @@ public class ActivationsTest {
 
     @Test
     void sigmoid_backward_matchesFiniteDifference() {
-        Tensor x = Tensor.from2D(new double[][]{{-1.2, 0.0, 2.3}});
+        Tensor x = Tensor.from2D(new float[][]{{-1.2f, 0.0f, 2.3f}});
 
         // analytic: d/dx sum(sigmoid(x)) = sigmoid'(x)
         Sigmoid s = new Sigmoid();
@@ -104,12 +104,12 @@ public class ActivationsTest {
             return ss.forward(t);
         }, x, 1e-3f);
 
-        TestSupport.assertTensorAllClose(analytic, numeric, 1e-3);
+        TestSupport.assertTensorAllClose(analytic, numeric, 1e-3f);
     }
 
     @Test
     void tanh_backward_matchesFiniteDifference() {
-        Tensor x = Tensor.from2D(new double[][]{{-0.7, 0.2, 1.1}});
+        Tensor x = Tensor.from2D(new float[][]{{-0.7f, 0.2f, 1.1f}});
 
         Tanh t = new Tanh();
         Tensor y = t.forward(x);
@@ -120,12 +120,12 @@ public class ActivationsTest {
             return tt.forward(u);
         }, x, 1e-3f);
 
-        TestSupport.assertTensorAllClose(analytic, numeric, 1e-3);
+        TestSupport.assertTensorAllClose(analytic, numeric, 1e-3f);
     }
 
     @Test
     void gelu_backward_matchesFiniteDifference() {
-        Tensor x = Tensor.from2D(new double[][]{{-1.5, -0.2, 0.0, 0.4, 2.0}});
+        Tensor x = Tensor.from2D(new float[][]{{-1.5f, -0.2f, 0.0f, 0.4f, 2.0f}});
 
         GELU g = new GELU();
         Tensor y = g.forward(x);
@@ -137,19 +137,19 @@ public class ActivationsTest {
         }, x, 1e-3f);
 
         // GELU is approximate + exp/tanh internally -> slightly looser tolerance
-        TestSupport.assertTensorAllClose(analytic, numeric, 2e-3);
+        TestSupport.assertTensorAllClose(analytic, numeric, 2e-3f);
     }
 
     @Test
     void softmax_backward_matchesFiniteDifference_forDotObjective() {
-        Tensor logits = Tensor.from2D(new double[][]{
-                { 0.3, -1.2, 2.0 },
-                { 1.0,  0.5, -0.7 }
+        Tensor logits = Tensor.from2D(new float[][]{
+                { 0.3f, -1.2f, 2.0f },
+                { 1.0f,  0.5f, -0.7f }
         });
 
-        Tensor upstream = Tensor.from2D(new double[][]{
-                { 0.7, -0.2, 1.3 },
-                { -1.1, 0.4, 0.9 }
+        Tensor upstream = Tensor.from2D(new float[][]{
+                { 0.7f, -0.2f, 1.3f },
+                { -1.1f, 0.4f, 0.9f }
         });
 
         // analytic: grad = softmax.backward(upstream)
@@ -160,7 +160,7 @@ public class ActivationsTest {
         // numeric: objective = sum( softmax(logits) * upstream )
         Tensor numeric = finiteDiffGradScalarObjective(x -> softmaxDotObjective(x, upstream), logits, 1e-3f);
 
-        TestSupport.assertTensorAllClose(analytic, numeric, 2e-3);
+        TestSupport.assertTensorAllClose(analytic, numeric, 2e-3f);
     }
 
     /** Numerical grad of objective = sum(f(x)) */
@@ -169,49 +169,49 @@ public class ActivationsTest {
 
         for (int r = 0; r < x.rows; r++) {
             for (int c = 0; c < x.cols; c++) {
-                double old = x.get(r, c);
+                float old = x.get(r, c);
 
                 x.set(r, c, old + eps);
-                double plus = f.apply(x).sum();
+                float plus = f.apply(x).sum();
 
                 x.set(r, c, old - eps);
-                double minus = f.apply(x).sum();
+                float minus = f.apply(x).sum();
 
                 x.set(r, c, old);
 
-                grad.set(r, c, (plus - minus) / (2.0 * eps));
+                grad.set(r, c, (plus - minus) / (2.0f * eps));
             }
         }
         return grad;
     }
 
     /** Numerical grad of any scalar objective J(x) */
-    private static Tensor finiteDiffGradScalarObjective(Function<Tensor, Double> objective, Tensor x, float eps) {
+    private static Tensor finiteDiffGradScalarObjective(Function<Tensor, Float> objective, Tensor x, float eps) {
         Tensor grad = new Tensor(x.rows, x.cols);
 
         for (int r = 0; r < x.rows; r++) {
             for (int c = 0; c < x.cols; c++) {
-                double old = x.get(r, c);
+                float old = x.get(r, c);
 
                 x.set(r, c, old + eps);
-                double plus = objective.apply(x);
+                float plus = objective.apply(x);
 
                 x.set(r, c, old - eps);
-                double minus = objective.apply(x);
+                float minus = objective.apply(x);
 
                 x.set(r, c, old);
 
-                grad.set(r, c, (plus - minus) / (2.0 * eps));
+                grad.set(r, c, (plus - minus) / (2.0f * eps));
             }
         }
         return grad;
     }
 
-    private static double softmaxDotObjective(Tensor logits, Tensor upstream) {
+    private static float softmaxDotObjective(Tensor logits, Tensor upstream) {
         Softmax sm = new Softmax();
         Tensor p = sm.forward(logits);
 
-        double s = 0.0;
+        float s = 0.0f;
         for (int r = 0; r < p.rows; r++) {
             for (int c = 0; c < p.cols; c++) {
                 s += p.data[r * p.cols + c] * upstream.data[r * upstream.cols + c];
