+++ src/test/java/io/github/kirstenali/deepj/training/SupervisedTrainingTest.java
@@ -18,14 +18,14 @@ public class SupervisedTrainingTest {
     @Test
     void supervisedTrainer_runs_andLossDoesNotExplode_onTinyRegression() {
         // y = 2x on 4 samples
-        Tensor x = Tensor.from2D(new double[][]{{0},{1},{2},{3}});
-        Tensor y = Tensor.from2D(new double[][]{{0},{2},{4},{6}});
+        Tensor x = Tensor.from2D(new float[][]{{0},{1},{2},{3}});
+        Tensor y = Tensor.from2D(new float[][]{{0},{2},{4},{6}});
 
         Linear model = new Linear(1, 1, new Random(1));
         Trainer trainer = SupervisedTraining.trainer(
                 model,
                 new MSELoss(),
-                AdamW.defaultAdamW(0.05),
+                AdamW.defaultAdamW(0.05f),
                 x, y,
                 123
         );
@@ -35,13 +35,13 @@ public class SupervisedTrainingTest {
 
         Assertions.assertTrue(Double.isFinite(l0));
         Assertions.assertTrue(Double.isFinite(l1));
-        Assertions.assertTrue(l1 <= l0 * 1.1, "loss should not blow up after one step");
+        Assertions.assertTrue(l1 <= l0 * 1.1f, "loss should not blow up after one step");
     }
 
     @Test
     void supervisedTrainer_keepsSampledInputTargetPairsAligned() {
-        Tensor x = Tensor.from2D(new double[][]{{0}, {1}, {2}, {3}, {4}});
-        Tensor y = Tensor.from2D(new double[][]{{0}, {1}, {2}, {3}, {4}});
+        Tensor x = Tensor.from2D(new float[][]{{0}, {1}, {2}, {3}, {4}});
+        Tensor y = Tensor.from2D(new float[][]{{0}, {1}, {2}, {3}, {4}});
 
         Layer identity = new Layer() {
             @Override
@@ -57,13 +57,13 @@ public class SupervisedTrainingTest {
         Trainer trainer = SupervisedTraining.trainer(
                 identity,
                 new MSELoss(),
-                AdamW.defaultAdamW(0.01),
+                AdamW.defaultAdamW(0.01f),
                 x, y,
                 123L
         );
 
         for (int i = 0; i < 20; i++) {
-            Assertions.assertEquals(0.0, trainer.trainStep(1), 1e-12,
+            Assertions.assertEquals(0.0f, trainer.trainStep(1), 1e-12f,
                     "aligned x/y sampling should keep identity-model loss at zero");
         }
     }
